requirements.txt
gradio
torch
numpy
tqdm
transformers
pytorch-lightning
pillow
matplotlib